llm_assistant/requirements.txt
streamlit>=1.35.0
anthropic>=0.28.0
psycopg2-binary>=2.9.9
pandas>=2.2.0
python-dotenv>=1.0.0
plotly>=5.22.0
